+++ Cesium3DTilesContent/src/BinaryToGltfConverter.cpp
@@ -36,7 +36,26 @@ CesiumAsync::Future<GltfConverterResult> BinaryToGltfConverter::convert(
     const std::span<const std::byte>& gltfBinary,
     const CesiumGltfReader::GltfReaderOptions& options,
     const AssetFetcher& assetFetcher) {
-  return assetFetcher.asyncSystem.createResolvedFuture(
-      convertImmediate(gltfBinary, options, assetFetcher));
+  return _gltfReader
+      .readGltfAndExternalData(
+          gltfBinary,
+          assetFetcher.asyncSystem,
+          assetFetcher.requestHeaders,
+          assetFetcher.pAssetAccessor,
+          assetFetcher.baseUrl,
+          options)
+      .thenInWorkerThread(
+          [&assetFetcher](CesiumGltfReader::GltfReaderResult&& gltfResult) {
+            if (gltfResult.model) {
+              gltfResult.model->extras["gltfUpAxis"] =
+                  static_cast<std::underlying_type_t<CesiumGeometry::Axis>>(
+                      assetFetcher.upAxis);
+            }
+            GltfConverterResult result;
+            result.model = std::move(gltfResult.model);
+            result.errors.errors = std::move(gltfResult.errors);
+            result.errors.warnings = std::move(gltfResult.warnings);
+            return result;
+          });
 }
 } // namespace Cesium3DTilesContent